Add more optimizations for reading/writing from Region files
Two new patches are included here, one for making use of memory-mapped
files and the other for reducing the number of checks against a world's
session lock. These can both bring pretty significant improvements to
the amount of CPU time spent in I/O while reading/writing chunks.

Unfortunately, due to the nature of the Region file format (primarily
being that the backing file can be dynamically re-sized), it is not
possible to memory-map the entire file. However, we can still memory-map
the entire offset and timestamp tables as they're part of a fixed
header.