new: Fast caching for biome layer sampling (#71)
This implements a much faster cache implementation for biome layer sampling. The default one is hugely flawed in being totally wrapped in a synchronized block, as well as maintaining its size by removing entries one-by-one from a linked list.

This implementation alternatively works on a fixed-size "lossy" cache which evicts old entries simply by overriding them. This hugely reduces the complexity of caching, creating a substantial improvement in sampling cost. This is especially relevant for stronghold generation, which takes a huge sample area to find a spawn location. It should generally improve the performance of chunk generation too, though.

To avoid using a synchronized block, we also keep a ThreadLocal instance of the biome layer stack.

Furthermore, in places where a vanilla layer is used from multiple other layers, the whole stack is duplicated. This is bad for caching, because it means nothing can be shared. This PR also ensures that those layers don't get duplicated for better caching and memory usage.

Also thanks to @SuperCoder7979 & @Kroppeb :)

---

As for the performance increase with the changes, on the seed 6250325407383374817 searching for a Mushroom Fields biome through /locatebiome minecraft:mushroom_fields takes less than a second (~5815 blocks away) and without these changes it takes about 3 seconds on the same seed. We'll get some better profiling stats soon.

---

Co-authored-by: SuperCoder7979 <k.pranav@gmail.com>
Co-authored-by: JellySquid <jellysquid3@users.noreply.github.com>