new: Faster implementation of ScaleLayer (#61)
This PR replaces ScaleLayer's sample method with a faster one that avoid sampling the surrounding biomes unless it is absolutely necessary, as ScaleLayer is the most used layer in the vanilla layer stack. This optimization is not that big when it comes to world generation, but it does have a noticeable impact when the locatebiome command is used. It also optimizes x >> 1 << 1 into x & ~1 for a bit smarter bitwise operation. There are no worldgen differences with this enabled.

With thanks to @gegy1000 and @Kroppeb

Co-authored-by: Gegy <gegy1000@gmail.com>
Co-authored-by: Kroppeb <kroppeb@users.noreply.github.com>
Co-authored-by: JellySquid <jellysquid3@users.noreply.github.com>