new: Avoid sampling in the chunk generator when possible (#64)
For some context, Minecraft uses two different perlin noise samplers to generate minecraft's terrain. The end result is the value of those two samplers, interpolated using another sampler. This interpolation sampler is normalized poorly, and it spends most of it's time at either 1 or 0.

Because the value is at the extremes, the interpolation used between the two perlin noise samplers is unnecessary and can be avoided unless the interpolation noise is within the [0, 1] range. In pseudocode, we are doing this:

```
if (interpolation >= 1) {
    return getNoise2();
} else if (interpolation <= 0) {
    return getNoise1();
} else {
    return lerp(getNoise1(), getNoise2(), interpolation);
}
```

Some rudimentary profiling shows that this causes 33% less sampling and reduces the overhead of the chunk generator. It also preserves the worldgen shape and doesn't cause any worldgen inconsistencies from some extensive testing.

Thanks again for @gegy1000 and @Kroppeb for helping write this and find bugs in my code smile