Correctly clean counter to fix contains method
The return value of "addTo" method returns "the old value, or the defaultreturn value if no value was present for the given key." so it has to check for 1 instead of 0 to find now empty counters. This fixes that the "containsKey" method returns true, even iwhen the counter is at 0.