new: Implement early-exit route for path-finding around dangerous neighbor blocks
Scanning for neighboring dangers around every candidate path finding node
is a major source of slowness in path-finding and accounts for ~90% of the
block reads.

This patch implements a cache which maintains whether or not a given chunk
section is free of such dangers. Path-finding code with it firstly tries to
determine whether or not all neighbors are contained in same chunk section,
and then retrieves or computes whether or not the block palette contains
dangerous blocks.

While the initial scan through the block palette is particularly slow, it
still greatly accelerates the vast majority of cases which hit the cache by
allowing us to early-exit when scanning for neighboring dangers since we can
know ahead of time that a chunk is devoid of them.

The cache must be enabled and disabled in controlled code paths as to avoid
world updates not being reflected in the path-finding cache. Empty chunk
sections are always assumed to devoid of any dangers.

In my simple benchmarks, this produced a ~50% reduction in server tick times
with 300 bees in a world, but it isn't specific to any one entity.