fix: block.hopper: hopper not updating after external modification of "shadow items"
Shadow items are item stacks that are in multiple inventories at the same time (requires obscure bug usage). Changing the stack in one inventory will modify the stack in another inventory, because it is the same mutable object being in both inventories. Therefore, caches relying on the item stacks inside the inventory being only accessed through the inventory are not invalidated correctly.

The fix circumvents this issue by disabling the caches for all inventories that at some point in their lifetime contained such a "shadow item".